+++ MoneyBuddi/src/dao/ICategoryDAO.java
@@ -13,9 +13,9 @@ public interface ICategoryDAO {
 	
 	void addCategory(Category category) throws SQLException;
 	
-	void deleteCategory(int id) throws SQLException; //User can only delete categories that he created and not the default ones
+	void deleteCategory(long id) throws SQLException; //User can only delete categories that he created and not the default ones
 	
-	Category getCategoryByID(int id) throws SQLException,InvalidDataException;
+	Category getCategoryByID(long id) throws SQLException,InvalidDataException;
 	
 	Collection<Category> getAllCategoriesByUser(User user) throws SQLException,InvalidDataException;
