+++ MoneyBuddi/src/dao/ICurrencyDAO.java
@@ -9,7 +9,7 @@ import model.Currency.CurrencyType;
 
 public interface ICurrencyDAO {
 	
-	Currency getCurrencyById(int id) throws SQLException;
+	Currency getCurrencyById(long id) throws SQLException;
 	
 	Currency getCurrencyByType(CurrencyType type) throws SQLException;
