+++ MoneyBuddi/src/dao/IUserDao.java
@@ -9,7 +9,7 @@ public interface IUserDao {
 	void saveUser(User u) throws SQLException;
 	void deleteUser(User u) throws SQLException;
 	void updateUser(User u) throws SQLException;
-	User getUserById(int id) throws SQLException;
+	User getUserById(long id) throws SQLException;
 	User getUserByUsername(String username) throws SQLException;
 	boolean checkIfEmailExists(String email) throws SQLException;
 	boolean validate(String email, String pass) throws SQLException;
